+++ docs/changelog.rst
@@ -3,6 +3,12 @@
 Changelog
 =========
 
+v0.5.0 (in development)
+-----------------------
+- Attachments' filenames can now be `None`
+- Added a `decompose()` function for decomposing
+  `~email.message.EmailMessage`\s into `MailItem`\s plus headers
+
 v0.4.0 (2021-03-13)
 -------------------
 - Using ``|``, ``&``, or ``^`` on a `MailItem` and a `str` now automatically